// src/types.ts
export type Stock = {
    approved: boolean;
    author: string;
    category: number;
    collection: string;
    description: string;
    main_file: string;
    keywords: string[];
    matured_content: boolean;
    id: number;
    type: string;
    created_at: string;
    updated_at: string;
    stock_id: string;
}